% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tdc.R
\name{style_clean}
\alias{style_clean}
\title{Create a new ggplot}
\usage{
style_clean(..., zero_line = "#344055", legend = FALSE)
}
\arguments{
\item{...}{arguments intended for a ggplot}

\item{zero_line}{a HEX colour value for the horizontal line (NA for no line)}

\item{legend}{a boolean value indicating whether to include the legend}
}
\description{
This function masks `ggplot()` from the ggplot2 package in order to apply additional
styles appropriate to The Data Collective charts
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

# example scatterplot
ggplot(iris, aes(Sepal.Length, Petal.Length, colour = as.factor(Species))) +
  geom_point() +
  labs(title = "Fisher's *Iris* data set",
       subtitle = "Demonstrating the TDC theme for ggplot2",
       x = "The x axis", y = "The y axis",
       caption = "Fisher's **Iris** data set") +
       style_tdc()
}
}
