% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{tdc_pal_con}
\alias{tdc_pal_con}
\title{The Data Collective palette with ramped colours}
\usage{
tdc_pal_con(palette = "tdc", alpha = 1, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose from 'tdc_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
The Data Collective palette with ramped colours
}
\examples{
library(scales)
show_col(tdc_pal()(10))

filled.contour(volcano,color.palette = tdc_pal(), asp=1)

}
