% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tdc.R
\name{theme_clean}
\alias{theme_clean}
\title{Theme Clean}
\usage{
theme_clean(
  base_family = "Roboto Condensed",
  base_size = 12,
  base_colour = "#1C1C1C",
  plot_title_family = base_family,
  plot_title_size = base_size * 2.5,
  plot_title_colour = base_colour,
  subtitle_family = "Roboto Thin",
  subtitle_size = base_size,
  subtiitle_colour = base_colour,
  title_family = base_family,
  title_size = base_size * 1.2,
  title_colour = base_colour,
  caption_family = subtitle_family,
  caption_size = base_size * 0.8,
  caption_colour = base_colour,
  panel_spacing = 25,
  grid_y = 1,
  grid_x = 1,
  grid_colour = "#ECECEC"
)
}
\arguments{
\item{base_family, base_size, base_colour}{base font family, size and colour}

\item{plot_title_family, plot_title_size, plot_title_colour}{plot title font family, size and colour}

\item{subtitle_family, subtitle_size, subtitle_colour}{subtitle font family size, and colour}

\item{title_family, title_size, title_colour}{title font family, size and colour}

\item{caption_family, caption_size, caption_colour}{caption font family, size and colour}

\item{panel_spacing}{the gap between panels in points}

\item{grid_y, grid_x}{the linestyle numeric 0-8 for the x/y gridlines}

\item{grid_colour}{the colour of the grid lines}
}
\description{
A consistent ggplot theme for use by The Data Collective
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

# example scatterplot
ggplot(iris, aes(Sepal.Length, Petal.Length, colour = as.factor(Species))) +
  geom_hline(yintercept = 0, colour = "#344055", size = 0.5) +
  geom_point() +
  labs(title = "Fisher's *Iris* data set",
       subtitle = "Demonstrating the TDC theme for ggplot2",
       x = "The x axis", y = "The y axis",
       caption = "Fisher's **Iris** data set") +
  facet_wrap(~ Species) +
  theme_clean()
}
}
