% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tdc.R
\name{theme_tdc}
\alias{theme_tdc}
\title{Theme TDC}
\usage{
theme_tdc(
  base_family = "Roboto Condensed",
  base_size = 12,
  plot_title_family = base_family,
  plot_title_size = 28,
  subtitle_family = "Roboto Condensed",
  subtitle_size = 12,
  title_family = base_family,
  title_size = 14,
  caption_family = base_family,
  caption_size = 10
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_size}{plot title font family and size}

\item{subtitle_family, subtitle_size}{subtitle font family and size}

\item{title_family, title_size}{subtitle font family and size}

\item{caption_family, caption_size}{caption font family and size}
}
\description{
A consistent ggplot theme for use by The Data Collective
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

# example scatterplot
ggplot(iris, aes(Sepal.Length, Petal.Length, colour = as.factor(Species))) +
  geom_hline(yintercept = 0, colour = "#344055", size = 0.5) +
  geom_point() +
  labs(title = "Fisher's *Iris* data set",
       subtitle = "Demonstrating the TDC theme for ggplot2",
       x = "The x axis", y = "The y axis",
       caption = "Fisher's **Iris** data set") +
  facet_wrap(~ Species) +
  theme_tdc()
}
}
